+++ src/Models/Orderrows/Collection/Orderrows.php
@@ -186,7 +186,7 @@ class Orderrows implements Model
         return $this;
     }
 
-    protected function convertFromData(string $key, $value)
+    protected function convertFromArrayData(string $key, $value)
     {
         if ($key === 'items') {
             $items = [];