+++ src/Models/Orderrows/Resource/Orderrows.php
@@ -906,7 +906,7 @@ class Orderrows implements Model
         return $this;
     }
 
-    protected function convertFromData(string $key, $value)
+    protected function convertFromArrayData(string $key, $value)
     {
         if ($key === 'attributes') {
             $items = [];