+++ src/Models/Orders/Collection/Orders.php
@@ -186,7 +186,7 @@ class Orders implements Model
         return $this;
     }
 
-    protected function convertFromData(string $key, $value)
+    protected function convertFromArrayData(string $key, $value)
     {
         if ($key === 'items') {
             $items = [];