+++ src/Models/Orders/Orders/Patch.php
@@ -594,7 +594,7 @@ class Patch implements Model
         return $this;
     }
 
-    protected function convertFromData(string $key, $value)
+    protected function convertFromArrayData(string $key, $value)
     {
         if ($key === 'orderrows') {
             $items = [];