+++ src/Models/Productattributevalues/Collection/Productattributevalues.php
@@ -186,7 +186,7 @@ class Productattributevalues implements Model
         return $this;
     }
 
-    protected function convertFromData(string $key, $value)
+    protected function convertFromArrayData(string $key, $value)
     {
         if ($key === 'items') {
             $items = [];