+++ src/Models/Productphotos/Productphotos/Put.php
@@ -42,7 +42,7 @@ class Put implements Model
         return $this;
     }
 
-    protected function convertFromData(string $key, $value)
+    protected function convertFromArrayData(string $key, $value)
     {
         if ($key === 'productphotos') {
             $items = [];