+++ src/Models/Products/Child/Resource/Productattributesets.php
@@ -66,7 +66,7 @@ class Productattributesets implements Model
         return $this;
     }
 
-    protected function convertFromData(string $key, $value)
+    protected function convertFromArrayData(string $key, $value)
     {
         if ($key === 'collection') {
             $items = [];