+++ src/Models/Products/Child/Resource/Productshippingcosts.php
@@ -138,7 +138,7 @@ class Productshippingcosts implements Model
         return $this;
     }
 
-    protected function convertFromData(string $key, $value)
+    protected function convertFromArrayData(string $key, $value)
     {
         if ($key === 'collection') {
             $items = [];