+++ src/Models/Products/Collection/Products.php
@@ -186,7 +186,7 @@ class Products implements Model
         return $this;
     }
 
-    protected function convertFromData(string $key, $value)
+    protected function convertFromArrayData(string $key, $value)
     {
         if ($key === 'items') {
             $items = [];