+++ src/Models/Products/Resource/Attributecombinations.php
@@ -474,7 +474,7 @@ class Attributecombinations implements Model
         return $this;
     }
 
-    protected function convertFromData(string $key, $value)
+    protected function convertFromArrayData(string $key, $value)
     {
         if ($key === 'combination') {
             $items = [];