+++ src/Models/Products/Resource/Productattributesets.php
@@ -162,7 +162,7 @@ class Productattributesets implements Model
         return $this;
     }
 
-    protected function convertFromData(string $key, $value)
+    protected function convertFromArrayData(string $key, $value)
     {
         if ($key === 'items') {
             $items = [];