+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## v0.12.0
 
 - Moved generator code to own namespace
-- Removed onlyFilledProperties parameter from endpoints and added `onlyFilledPropertiesInToArrayData(false)` to the ToArray trait.
-- Renamed protected methods of ToArray
+- Renamed methods of the `ToArray` trait
     - `removeFromRequestData(string $key): bool` -> `removeFromToArrayData(string $key): bool`
     - `covertToData(string $key, $value)` -> `convertToArrayData(string $key, $value)`
 - Removed protected properties of ToArray trait in `toArray()` result.
-- 
+- Renamed methods of the `FromArray` trait
+    - `convertFromData(string $key, $value)` -> `convertFromArrayData(string $key, $value)`
+